linting and cleaning up imports
